+++ CHANGELOG.md
@@ -3,6 +3,7 @@ All notable changes to this project will be documented in this file.
 
 ## [0.7.3] - 2018-11-25
 - Added DPS/pDPS/eDPS to weapon-tooltips
+- Added a column that shows which players has an item in the group summary table 
 - Fixed a bug where snapshots didnt get triggered on zone-change
 - Fixed a bug where you could select all stash-tabs (max is 20)
 - Minor style-tweaks